requirements.txt
certifi==2018.1.18
chardet==3.0.4
google-api-python-client==1.6.5
gspread==0.6.2
httplib2==0.10.3
idna==2.6
oauth2client==4.1.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyDrive==1.3.1
PyYAML==3.12
requests==2.18.4
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
